adds ContactAnchor and ContactDialog
