+++ src/components/ContactDialog/index.tsx
@@ -1,48 +1,82 @@
-import React from "react";
-import { Dialog, DialogTitle, Button } from "@mui/material";
-import ContactDlg from "../ContactDlg";
-// export interface SimpleDialogProps {
-//   open: boolean;
-//   selectedValue: string;
-//   onClose: (value: string) => void;
-// }
+import { Dialog, DialogActions, DialogTitle, Button } from "@mui/material";
+import Grid from "@mui/material/Grid2";
+import TextField from "@mui/material/TextField";
+import FormHelperText from "@mui/material/FormHelperText";
 
-// function SimpleDialog(props: SimpleDialogProps) {
-//   const handleClose = () => {
-//     props.onClose(props.selectedValue);
-//   };
+export interface ContactDialogProps {
+  open: boolean;
+  selectedValue: string;
+  onClose: (value: string) => void;
+}
 
-//   return (
-//     <>
-//       <Dialog onClose={props.onClose} open={props.open}>
-//         <DialogTitle>Foprksalot</DialogTitle>
-//       </Dialog>
-//     </>
-//   );
-// }
-
-export default function ContactDialog() {
-  const [open, setOpen] = React.useState(false);
-
-  //   toggling dialog show/hide
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
+export default function ContactDialog(props: ContactDialogProps) {
   const handleClose = () => {
-    setOpen(false);
+    props.onClose(props.selectedValue);
   };
 
   return (
     <>
-      <Button variant="outlined" onClick={handleClickOpen}>
-        Contact Us
-      </Button>
-      <ContactDlg
-        open={open}
-        onClose={handleClose}
-        selectedValue="flog"
-      ></ContactDlg>
+      <Dialog onClose={props.onClose} open={props.open}>
+        <DialogTitle>We would love to hear from you</DialogTitle>
+        <Grid container spacing={2} padding={4}>
+          <Grid container spacing={3} padding={2}>
+            <Grid>
+              <TextField
+                id="first-name"
+                label="First Name"
+                required
+                variant="standard"
+                fullWidth
+              />
+            </Grid>
+            <Grid>
+              <TextField
+                id="last-name"
+                label="Last Name"
+                required
+                variant="standard"
+                fullWidth
+              />
+            </Grid>
+          </Grid>
+          <Grid padding={2}>
+            <TextField
+              id="email"
+              label="Email"
+              required
+              fullWidth
+              variant="standard"
+              aria-describedby="email-address-helper-text"
+            />
+            <FormHelperText id="email-address-helper-text">
+              We'll never share your email address
+            </FormHelperText>
+
+            <TextField
+              id="email-subject"
+              label="Subject"
+              variant="standard"
+              fullWidth
+            />
+
+            <TextField
+              id="message"
+              label=""
+              required
+              multiline
+              rows={3}
+              fullWidth
+              variant="standard"
+              placeholder="Message"
+              sx={{ marginTop: 2 }} // Add spacing above
+            />
+          </Grid>
+        </Grid>
+        <DialogActions>
+          <Button onClick={handleClose}>Cancel</Button>
+          <Button type="submit">Send Message</Button>
+        </DialogActions>
+      </Dialog>
     </>
   );
 }
